+++ zkos/recovery/src/contracts.rs
@@ -1,0 +1,138 @@
+use alloy::sol;
+
+sol! {
+    #[derive(Debug)]
+    function commitBatchesSharedBridge(
+        address _chainAddress,
+        uint256 _processFrom,
+        uint256 _processTo,
+        bytes _commitData
+    );
+    event BlockCommit(uint256 indexed batchNumber, bytes32 indexed batchHash, bytes32 indexed commitment);
+
+
+    struct L2CanonicalTransaction {
+        uint256 txType;
+        uint256 from;
+        uint256 to;
+        uint256 gasLimit;
+        uint256 gasPerPubdataByteLimit;
+        uint256 maxFeePerGas;
+        uint256 maxPriorityFeePerGas;
+        uint256 paymaster;
+        uint256 nonce;
+        uint256 value;
+        // In the future, we might want to add some
+        // new fields to the struct. The `txData` struct
+        // is to be passed to account and any changes to its structure
+        // would mean a breaking change to these accounts. To prevent this,
+        // we should keep some fields as "reserved"
+        // It is also recommended that their length is fixed, since
+        // it would allow easier proof integration (in case we will need
+        // some special circuit for preprocessing transactions)
+        uint256[4] reserved;
+        bytes data;
+        bytes signature;
+        uint256[] factoryDeps;
+        bytes paymasterInput;
+        // Reserved dynamic type for the future use-case. Using it should be avoided,
+        // But it is still here, just in case we want to enable some additional functionality
+        bytes reservedDynamic;
+    }
+
+    event NewPriorityRequest(
+        uint256 txId,
+        bytes32 txHash,
+        uint64 expirationTimestamp,
+        L2CanonicalTransaction transaction,
+        bytes[] factoryDeps
+    );
+
+    event GenesisUpgrade(
+        address indexed _zkChain,
+        L2CanonicalTransaction _l2Transaction,
+        uint256 indexed _protocolVersion,
+        bytes[] _factoryDeps
+    );
+
+    #[derive(Debug)]
+    struct StoredBatchInfo {
+        uint64 batchNumber;
+        bytes32 batchHash;
+        uint64 indexRepeatedStorageChanges;
+        uint256 numberOfLayer1Txs;
+        bytes32 priorityOperationsHash;
+        bytes32 dependencyRootsRollingHash;
+        bytes32 l2LogsTreeRoot;
+        uint256 timestamp;
+        bytes32 commitment;
+    }
+
+    #[derive(Debug)]
+    struct CommitBatchInfoZKsyncOS {
+        uint64 batchNumber;
+        bytes32 newStateCommitment;
+        uint256 numberOfLayer1Txs;
+        bytes32 priorityOperationsHash;
+        bytes32 dependencyRootsRollingHash;
+        bytes32 l2LogsTreeRoot;
+        address l2DaValidator;
+        bytes32 daCommitment;
+        uint64 firstBlockTimestamp;
+        uint64 lastBlockTimestamp;
+        uint256 chainId;
+        bytes operatorDAInput;
+    }
+
+    // A dummy function that takes the same parameters we encoded.
+    function __decodeParams(StoredBatchInfo, CommitBatchInfoZKsyncOS[]);
+
+
+    // Stuff needed for genesis upgrade tx decoding
+    function upgrade(address delegateTo, bytes _calldata);
+    function genesisUpgrade(
+        bool _isZKsyncOS,
+        uint256 _chainId,
+        address _ctmDeployer,
+        bytes calldata _fixedForceDeploymentsData,
+        bytes calldata _additionalForceDeploymentsData
+    );
+
+    #[derive(Debug)]
+    struct FixedForceDeploymentsData {
+        uint256 l1ChainId;
+        uint256 eraChainId;
+        address l1AssetRouter;
+        bytes32 l2TokenProxyBytecodeHash;
+        address aliasedL1Governance;
+        uint256 maxNumberOfZKChains;
+        bytes bridgehubBytecodeInfo;
+        bytes l2AssetRouterBytecodeInfo;
+        bytes l2NtvBytecodeInfo;
+        bytes messageRootBytecodeInfo;
+        bytes chainAssetHandlerBytecodeInfo;
+        bytes beaconDeployerInfo;
+        address l2SharedBridgeLegacyImpl;
+        address l2BridgedStandardERC20Impl;
+        // The forced beacon address. It is needed only for internal testing.
+        // MUST be equal to 0 in production.
+        // It will be the job of the governance to ensure that this value is set correctly.
+        address dangerousTestOnlyForcedBeacon;
+    }
+
+
+    #[derive(Debug)]
+    struct ZKChainSpecificForceDeploymentsData {
+        bytes32 baseTokenAssetId;
+        address l2LegacySharedBridge;
+        address predeployedL2WethAddress;
+        address baseTokenL1Address;
+        /// @dev Some info about the base token, it is
+        /// needed to deploy weth token in case it is not present
+        string baseTokenName;
+        string baseTokenSymbol;
+    }
+
+
+
+}
